requirements.txt
asgiref==3.8.1
Django==4.2.13
sqlparse==0.5.0
typing_extensions==4.12.0
django-ninja
stellar-sdk
python-dotenv